Refactored the code to be more modular.
The main file has been split into sub files based on functionality, this allows for easier readability and maintenance of the code.
The code has been split into the following modules:
- database.py
- scheme.py
- tokenizer.py
- learner.py
- cache.py
- ui.py
- logger.py

These modules now contain the functions that were previously in the main file.
The main file has been updated to import from these modules.
Due to environment issues, tests were unable to be run, however all changes have been made.
